mozzi_midi as header only
